Rename notes to pynotes and move things around
